Added description to cli functions
Added utils logger

Fixed settings.json
